feat: add share tooltip

+++ src/modules/workshop/components/Modal/SetProjectNameModal/index.tsx
@@ -61,7 +61,7 @@ const SetProjectNameModal = ({ type }: Props) => {
     }
 
     const res = await saveProject(payload)
-    if (res !== 'failed') {
+    if (!!res && res !== 'failed') {
       if (type === 'save-exit') {
         createProject()
         deleteAlls()
